requirements.txt
tensorflow
pandas
numpy
matplotlib
fastapi
uvicorn
pillow